Module-13-Model-Context-Protocol/RAG_MCP_Chroma_Gradio/requirements.txt
openai
python-dotenv
gradio
httpx
fastmcp
PyMuPDF
langchain-chroma
PyMuPDF 
tiktoken 
openai 
psycopg2-binary